+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 1.1.0 (not yet released)
 * (base) Make JwtUser extendable (#20)
 * (base) Make rest endpoints configurable (#19)
+* (base) Mark JwtTokenService.getExpiration as deprecated
+* (base) Mark RefreshTokenStore.getRefreshExpiration as deprecated
 
 ### 1.0.0 (2017-09-05)
 * (all) Add spring configuration properties support for IDE auto-completion